pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "makegis"
version = "0.1.0"
description = "Build spatial databases"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [{name = "Chris Thiange"}]

dependencies = ["psycopg[binary]", "python-dotenv", "pyyaml"]

[project.scripts]
mkgs = "makegis.cli:cli"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]